requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
cachetools==4.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.20
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
easydict==1.9
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
idna==2.10
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.2
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.0
protobuf==3.11.3
pycocotools==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.0
wrapt==1.12.1
